AIRS/dags/model_retrain.py
"""
Model retraining DAG for AIRS.

Orchestrates periodic model retraining with walk-forward validation.
"""

from datetime import datetime, timedelta
from typing import Any

from airflow import DAG
from airflow.operators.python import PythonOperator, BranchPythonOperator
from airflow.operators.empty import EmptyOperator
from airflow.utils.task_group import TaskGroup


# Default arguments
default_args = {
    "owner": "airs",
    "depends_on_past": False,
    "email": ["ml-alerts@example.com"],
    "email_on_failure": True,
    "email_on_retry": False,
    "retries": 2,
    "retry_delay": timedelta(minutes=10),
    "execution_timeout": timedelta(hours=4),
}


def check_retrain_needed(**context) -> str:
    """Check if model retraining is needed."""
    # Criteria for retraining:
    # 1. Scheduled monthly retrain
    # 2. Performance degradation detected
    # 3. Significant data drift

    execution_date = context["execution_date"]

    # Check if it's the first Monday of the month (scheduled retrain)
    is_first_monday = execution_date.day <= 7 and execution_date.weekday() == 0

    # TODO: Check for performance degradation
    # from airs.monitoring.drift import check_model_performance
    # performance_degraded = check_model_performance()
    performance_degraded = False

    # TODO: Check for data drift
    # from airs.monitoring.drift import check_data_drift
    # data_drift_detected = check_data_drift()
    data_drift_detected = False

    if is_first_monday or performance_degraded or data_drift_detected:
        return "prepare_training_data"
    else:
        return "skip_retrain"


def prepare_training_data(**context) -> dict[str, Any]:
    """Prepare training data for model retraining."""
    from datetime import date

    # TODO: Load historical features and labels
    # from airs.db.repository import FeatureRepository, LabelRepository
    # feature_repo = FeatureRepository()
    # label_repo = LabelRepository()

    execution_date = context["execution_date"]

    # Use walk-forward approach: train on all data up to execution date
    train_end_date = execution_date - timedelta(days=1)  # Exclude today
    train_start_date = train_end_date - timedelta(days=5 * 365)  # 5 years of data

    # TODO: Load and prepare data
    # features = feature_repo.get_features(train_start_date, train_end_date)
    # labels = label_repo.get_labels(train_start_date, train_end_date)

    return {
        "train_start": train_start_date.strftime("%Y-%m-%d"),
        "train_end": train_end_date.strftime("%Y-%m-%d"),
        "n_samples": 1260,  # Mock value
        "n_features": 50,
    }


def train_baseline_models(**context) -> dict[str, Any]:
    """Train baseline models (LogReg, ThresholdModel)."""
    import mlflow

    ti = context["ti"]
    data_info = ti.xcom_pull(task_ids="prepare_training_data")

    # TODO: Load data and train models
    # from airs.models.baseline import LogisticModel, ThresholdModel

    with mlflow.start_run(run_name="baseline_models", nested=True):
        # Train logistic regression
        # logreg = LogisticModel()
        # logreg.fit(X_train, y_train)

        # Log metrics
        mlflow.log_metric("logreg_auc", 0.72)
        mlflow.log_metric("threshold_precision", 0.65)

    return {
        "logreg_auc": 0.72,
        "threshold_precision": 0.65,
    }


def train_xgboost(**context) -> dict[str, Any]:
    """Train XGBoost model."""
    import mlflow

    with mlflow.start_run(run_name="xgboost", nested=True):
        # TODO: Train XGBoost
        # from airs.models.tree_ensemble import XGBoostModel
        # model = XGBoostModel()
        # model.fit(X_train, y_train)

        mlflow.log_metric("xgb_auc", 0.78)
        mlflow.log_metric("xgb_precision", 0.70)

    return {
        "auc": 0.78,
        "precision": 0.70,
    }


def train_lightgbm(**context) -> dict[str, Any]:
    """Train LightGBM model."""
    import mlflow

    with mlflow.start_run(run_name="lightgbm", nested=True):
        # TODO: Train LightGBM
        # from airs.models.tree_ensemble import LightGBMModel
        # model = LightGBMModel()
        # model.fit(X_train, y_train)

        mlflow.log_metric("lgbm_auc", 0.79)
        mlflow.log_metric("lgbm_precision", 0.72)

    return {
        "auc": 0.79,
        "precision": 0.72,
    }


def train_random_forest(**context) -> dict[str, Any]:
    """Train Random Forest model."""
    import mlflow

    with mlflow.start_run(run_name="random_forest", nested=True):
        # TODO: Train Random Forest
        # from airs.models.tree_ensemble import RandomForestModel
        # model = RandomForestModel()
        # model.fit(X_train, y_train)

        mlflow.log_metric("rf_auc", 0.76)
        mlflow.log_metric("rf_precision", 0.68)

    return {
        "auc": 0.76,
        "precision": 0.68,
    }


def train_ensemble(**context) -> dict[str, Any]:
    """Train stacking ensemble."""
    import mlflow

    ti = context["ti"]

    # Get base model results
    xgb_result = ti.xcom_pull(task_ids="train_models.train_xgboost")
    lgbm_result = ti.xcom_pull(task_ids="train_models.train_lightgbm")
    rf_result = ti.xcom_pull(task_ids="train_models.train_random_forest")

    with mlflow.start_run(run_name="stacking_ensemble", nested=True):
        # TODO: Train ensemble
        # from airs.models.ensemble import StackingEnsemble
        # ensemble = StackingEnsemble(base_models=[xgb, lgbm, rf])
        # ensemble.fit(X_train, y_train)

        mlflow.log_metric("ensemble_auc", 0.82)
        mlflow.log_metric("ensemble_precision", 0.75)

    return {
        "auc": 0.82,
        "precision": 0.75,
    }


def run_validation(**context) -> dict[str, Any]:
    """Run walk-forward validation."""
    import mlflow

    ti = context["ti"]
    ensemble_result = ti.xcom_pull(task_ids="train_ensemble")

    # TODO: Run walk-forward validation
    # from airs.models.validation import WalkForwardValidator
    # validator = WalkForwardValidator()
    # results = validator.validate(ensemble, X, y)

    validation_results = {
        "avg_auc": 0.80,
        "avg_precision": 0.73,
        "avg_recall": 0.65,
        "std_auc": 0.03,
        "n_folds": 5,
    }

    with mlflow.start_run(run_name="validation", nested=True):
        for key, value in validation_results.items():
            mlflow.log_metric(key, value)

    return validation_results


def compare_to_production(**context) -> dict[str, Any]:
    """Compare new model to production model."""
    ti = context["ti"]
    validation_results = ti.xcom_pull(task_ids="run_validation")

    # TODO: Load production model metrics
    # from airs.db.repository import ModelRepository
    # prod_metrics = ModelRepository().get_production_metrics()

    production_auc = 0.78  # Mock production value

    new_auc = validation_results["avg_auc"]
    improvement = new_auc - production_auc

    return {
        "production_auc": production_auc,
        "new_auc": new_auc,
        "improvement": improvement,
        "should_deploy": improvement > 0.01,  # Deploy if >1% improvement
    }


def decide_deployment(**context) -> str:
    """Decide whether to deploy new model."""
    ti = context["ti"]
    comparison = ti.xcom_pull(task_ids="compare_to_production")

    if comparison["should_deploy"]:
        return "deploy_model"
    else:
        return "skip_deployment"


def deploy_model(**context) -> dict[str, Any]:
    """Deploy new model to production."""
    import mlflow

    # TODO: Register model in MLflow Model Registry
    # model_uri = f"runs:/{run_id}/model"
    # mlflow.register_model(model_uri, "airs-ensemble")

    # TODO: Transition to production
    # client = mlflow.tracking.MlflowClient()
    # client.transition_model_version_stage(
    #     name="airs-ensemble",
    #     version=new_version,
    #     stage="Production"
    # )

    return {
        "deployed": True,
        "model_version": "v2",  # Mock version
    }


def send_training_report(**context) -> dict[str, Any]:
    """Send training completion report."""
    ti = context["ti"]

    validation_results = ti.xcom_pull(task_ids="run_validation")
    comparison = ti.xcom_pull(task_ids="compare_to_production")

    # TODO: Send email report
    # from airs.utils.notifications import send_training_report
    # send_training_report(validation_results, comparison)

    return {
        "report_sent": True,
    }


# Create DAG
with DAG(
    dag_id="airs_model_retrain",
    default_args=default_args,
    description="Monthly model retraining with walk-forward validation",
    schedule_interval="0 6 * * 1",  # Every Monday at 6 AM UTC
    start_date=datetime(2024, 1, 1),
    catchup=False,
    tags=["airs", "ml", "training"],
    max_active_runs=1,
) as dag:

    # Start
    start = EmptyOperator(task_id="start")

    # Check if retraining is needed
    check_retrain = BranchPythonOperator(
        task_id="check_retrain_needed",
        python_callable=check_retrain_needed,
    )

    # Skip path
    skip_retrain = EmptyOperator(task_id="skip_retrain")

    # Prepare data
    prepare_data = PythonOperator(
        task_id="prepare_training_data",
        python_callable=prepare_training_data,
    )

    # Train base models
    with TaskGroup("train_models") as train_models:
        baseline = PythonOperator(
            task_id="train_baseline",
            python_callable=train_baseline_models,
        )

        xgboost = PythonOperator(
            task_id="train_xgboost",
            python_callable=train_xgboost,
        )

        lightgbm = PythonOperator(
            task_id="train_lightgbm",
            python_callable=train_lightgbm,
        )

        random_forest = PythonOperator(
            task_id="train_random_forest",
            python_callable=train_random_forest,
        )

    # Train ensemble
    ensemble = PythonOperator(
        task_id="train_ensemble",
        python_callable=train_ensemble,
    )

    # Validation
    validation = PythonOperator(
        task_id="run_validation",
        python_callable=run_validation,
    )

    # Compare to production
    compare = PythonOperator(
        task_id="compare_to_production",
        python_callable=compare_to_production,
    )

    # Deployment decision
    decide = BranchPythonOperator(
        task_id="decide_deployment",
        python_callable=decide_deployment,
    )

    # Deploy
    deploy = PythonOperator(
        task_id="deploy_model",
        python_callable=deploy_model,
    )

    # Skip deployment
    skip_deploy = EmptyOperator(task_id="skip_deployment")

    # Send report
    report = PythonOperator(
        task_id="send_training_report",
        python_callable=send_training_report,
        trigger_rule="none_failed_min_one_success",
    )

    # End
    end = EmptyOperator(
        task_id="end",
        trigger_rule="none_failed_min_one_success",
    )

    # Define dependencies
    start >> check_retrain
    check_retrain >> skip_retrain >> end
    check_retrain >> prepare_data >> train_models >> ensemble >> validation >> compare >> decide
    decide >> deploy >> report >> end
    decide >> skip_deploy >> report >> end
