AIRS/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airs"
version = "0.1.0"
description = "AI-Driven Early-Warning System for Portfolio Drawdown Risk"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "AIRS Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",

    # Data Sources
    "yfinance>=0.2.31",
    "fredapi>=0.5.1",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",

    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.12.0",

    # Machine Learning
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "shap>=0.43.0",
    "hmmlearn>=0.3.0",
    "imbalanced-learn>=0.11.0",

    # Deep Learning (optional)
    "torch>=2.0.0",

    # API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Experiment Tracking
    "mlflow>=2.9.0",

    # Configuration
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",

    # Utilities
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    "hypothesis>=6.92.0",
    "pandas-stubs>=2.0.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
]

airflow = [
    "apache-airflow>=2.7.0",
    "apache-airflow-providers-postgres>=5.0.0",
]

notebooks = [
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
]

all = [
    "airs[dev,airflow,notebooks]",
]

[project.scripts]
airs = "airs.cli:app"

[project.urls]
Homepage = "https://github.com/airs/airs"
Documentation = "https://github.com/airs/airs#readme"
Repository = "https://github.com/airs/airs"

[tool.setuptools.packages.find]
where = ["."]
include = ["airs*"]

[tool.setuptools.package-data]
airs = ["config/*.yaml"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.ruff_cache
    | \.venv
    | venv
    | build
    | dist
    | mlflow
)/
'''

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "build",
    "dist",
    "mlflow",
]

[tool.ruff.isort]
known-first-party = ["airs"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    "build",
    "dist",
    "mlflow",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=airs --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["airs"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
